+++ src/Document.js
@@ -1,3 +1,12 @@
+/*
+ * @fileoverview    Manages scanning, inserting and highlighting citations in
+ *                  the currently opened document in the editor.
+ * @author          Wasif Hasan Baig <pr.wasif@gmail.com>
+ * @copyright       Wasif Hasan Baig 2015
+ * @since           0.1.0
+ * @license         MIT
+ */
+
 /*jslint vars: true, nomen: true, plusplus: true */
 /*global define, brackets */
 define(function (require, exports, module) {
@@ -15,22 +24,28 @@ define(function (require, exports, module) {
     var C               = require("src/utils/Constants");
 
     function insertText(string) {
-        var editor = EditorManager.getCurrentFullEditor();
-        var doc = DocumentManager.getCurrentDocument();
-        var selections = editor.getSelections();
-        var edits = [];
+        var edits      = [];
+        var document   = DocumentManager.getCurrentDocument();
+        var selections = EditorManager.getCurrentFullEditor()
+                                      .getSelections();
 
         _.forEach(selections, function (sel) {
-            edits.push({edit: {text: string, start: sel.start, end: sel.end}});
+            edits.push({
+                edit : {
+                    text  : string,
+                    start : sel.start,
+                    end   : sel.end
+                }
+            });
         });
 
-        doc.doMultipleEdits(edits);
+        document.doMultipleEdits(edits);
     }
 
     function extractKeysFromText(fileContents) {
-        // console.log("SCAN: Extracting citation keys...")
-        var keys = [];
+        var keys  = [];
         var match = fileContents.match(C.REGEXP_MATCH_BRACKETS);
+
         _.forEach(match, function (citesEnclosedInBrackets) {
             var items = citesEnclosedInBrackets.match(C.REGEXP_CITE_KEYS);
             _.forEach(items, function (citeKey) {
@@ -38,20 +53,25 @@ define(function (require, exports, module) {
                 keys.push(citeKey.slice(1));
             });
         });
+
         return _.uniq(keys);
     }
 
     function indexKeysLocations(fileContents, keys) {
-        var lines = fileContents.split("\n");
-        var lineCount = 0; // 0-based line count
+        var lines         = fileContents.split("\n");
+        var lineCount     = 0; // 0-based line count
         var selectionKeys = {};
+
         _.forEach(lines, function (line) {
+
             _.forEach(keys, function (key) {
                 if (!selectionKeys[key]) {
                     selectionKeys[key] = [];
                 }
+
                 var re = new RegExp(key, "g");
                 var match;
+
                 while ((match = re.exec(line)) !== null) {
                     var obj = {
                         line    : lineCount,
@@ -61,105 +81,109 @@ define(function (require, exports, module) {
                     selectionKeys[key].push(obj);
                 }
             });
+
             lineCount++;
         });
+
         _.forEach(selectionKeys, function (array) {
             array[0].current = true;
         });
+
         return selectionKeys;
     }
 
-    function getEditor(editor) {
+    function _getEditor(editor) {
         // getting editor if no editor provided
         if (!editor) {
             editor = EditorManager.getCurrentFullEditor();
         }
+
         // if still no editor found, throwing error
         if (!editor) {
-            //            console.log("No Editor found!")
             return false;
         }
+
         return editor;
     }
 
-    //    function _updateDocumentPath(editor) {
-    //        this.path = editor.document.file._path
-    //        return this.path
-    //    }
-
     function getFileExtension(editor) {
-        return editor.document.file._path.split(".").slice(-1)[0];
+        return editor.document
+                     .file
+                     ._path
+                     .split(".")
+                     .slice(-1)[0];
     }
 
     function getNextKey(key) {
         /*jshint validthis: true */
-        var col = this.keysLocationHash[key];
-        var length = col.length;
-        var index = 0;
+
+        var col         = this.keysLocationHash[key];
+        var index       = 0;
+        var length      = col.length;
         var currentItem = null;
+
         _.forEach(col, function (item, idx) {
             if (!!item.current) {
                 index = idx;
                 currentItem = item;
             }
         });
+
         this.keysLocationHash[key][index].current = false;
+
         if (index === length - 1) {
             this.keysLocationHash[key][0].current = true;
         } else {
             this.keysLocationHash[key][index + 1].current = true;
         }
+
         return currentItem;
     }
 
-    function handleHighightCites(key) {
+    function _handleHighlighting(key) {
         /*jshint validthis: true */
+
         var editor = EditorManager.getCurrentFullEditor();
+
         if (!editor) {
             return false;
         }
+
         var curentCite = _.bind(getNextKey, this)(key.slice(1));
+
         editor.setCursorPos(curentCite.line, curentCite.index, true);
         editor.selectWordAt(editor.getSelection().start);
     }
 
-    function handleFileScan(editor) {
+    function _handleFileScanning(editor) {
         /*jshint validthis: true */
 
-        //        console.log("Scan Requested...")
+        editor = _getEditor(editor);
 
-        editor = getEditor(editor);
         if (!editor) {
             return false;
         }
 
         var fileExtension = getFileExtension(editor);
-        if (!_.contains(settings.docTypesToScanForCitationKeys, fileExtension)) {
-            // console.log("Scan Requested...Aborting: '" + fileExtension +
-            //             "' files not allowed to scan!");
+
+        if (!_.contains(settings.docTypesToScan, fileExtension)) {
             return false;
         }
 
-        var panelVisible = Channel.UI.request(Events.RQT_PANEL_VISIBILITY_STATUS);
-        //        console.log("PANEL-VISIBILITY", panelVisible)
-        if (!panelVisible) {
-            //            console.log("Scan Requested...Aborting: Panel is not visible!")
+        var isPanelVisible = Channel.UI.request(Events.RQT_PANEL_VISIBILITY_STATUS);
+
+        if (!isPanelVisible) {
             return false;
         }
 
-        //        var prevPath = this.path
-        //        var path = _.bind(_updateDocumentPath, this)(editor)
-        //        if (path == prevPath && this.scanned) {
-        //            console.log("Scan Requested...Aborting: Same scanned document!")
-        //            return false
-        //        }
-
-        //        console.log("Scan Requested...scanning!")
         var fileContents = editor.document.getText();
+
         if (!fileContents) {
             return [];
         }
+
         var keys = extractKeysFromText(fileContents);
+
         if (keys.length) {
             this.extractedKeys = keys;
             Channel.Zotero.trigger(Events.EVT_CITEKEYS_FOUND, keys);
@@ -170,19 +194,23 @@ define(function (require, exports, module) {
     }
 
     // insert the selected key into document at the current cursor position
-    function handleInserCitation() {
+    function _handleCiteInsertion() {
         var citeString = Channel.Zotero.request(Events.RQT_CITE_STRING);
+
         if (!citeString) {
             return;
         }
+
         insertText(citeString);
     }
 
-    function handleInsertBibliography() {
+    function _handleBiblioInsertion() {
         var biblioStringPromise = Channel.Zotero.request(Events.RQT_BIBLIOGRAPHY);
+
         if (!biblioStringPromise) {
             throw new Error("Error Inserting Bibliography in the Document");
         }
+
         biblioStringPromise.then(function (data) {
             insertText(data.result);
         });
@@ -190,22 +218,23 @@ define(function (require, exports, module) {
 
     function init() {
         /*jshint validthis: true */
-        Channel.Document.comply(Events.CMD_HIGHLIGHT_CITES, _.bind(handleHighightCites, this));
-        Channel.Extension.on(Events.EVT_PANEL_SHOWN, _.bind(handleFileScan, this));
-        Channel.Extension.on(Events.EVT_EDITOR_CHANGE, _.bind(handleFileScan, this));
+        Channel.Document.comply(Events.CMD_HIGHLIGHT_CITES, _.bind(_handleHighlighting, this));
+        Channel.Extension.on(Events.EVT_PANEL_SHOWN,        _.bind(_handleFileScanning, this));
+        Channel.Extension.on(Events.EVT_EDITOR_CHANGE,      _.bind(_handleFileScanning, this));
 
-        Channel.Extension.comply(C.CMD_ID_INSERT_CITE, handleInserCitation);
-        Channel.Extension.comply(C.CMD_ID_INSERT_BIBLIO, handleInsertBibliography);
+        Channel.Extension.comply(C.CMD_ID_INSERT_CITE,      _handleCiteInsertion);
+        Channel.Extension.comply(C.CMD_ID_INSERT_BIBLIO,    _handleBiblioInsertion);
     }
 
     function Document() {
         this.scanned = false;
         this.extractedKeys = [];
         this.keysLocationHash = {};
-        // this.path = ""
+
         Channel.Extension.on(Events.EVT_INIT, _.bind(init, this));
     }
 
+    // exporting the singleton Document object
     module.exports = new Document();
 
 });