+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The subsection are to be interpreted as defined [here](http://keepachangelog.com
 ### Changed
 - Updated gitignore.
 - Linted and style checked all the code.
+- `src/Document.js` cleaned up.
 
 ## [0.1.6] - 2015-03-19